-code cleanup

+++ app/src/main/java/com/example/bookapp/customViews/CommentDialog.java
@@ -13,8 +13,8 @@ import com.google.android.material.textfield.TextInputLayout;
 public class CommentDialog extends Dialog {
 
     private CommentDialogInterface commentDialogInterface;
-    private int postID;
-    public CommentDialog(@NonNull Context context,@NonNull CommentDialogInterface commentDialogInterface, int postID) {
+    private long postID;
+    public CommentDialog(@NonNull Context context,@NonNull CommentDialogInterface commentDialogInterface, long postID) {
         super(context);
         this.commentDialogInterface = commentDialogInterface;
         this.postID = postID;
@@ -35,6 +35,6 @@ public class CommentDialog extends Dialog {
     }
 
     public interface CommentDialogInterface {
-        void submitComment(String comment,int postID);
+        void submitComment(String comment,long postID);
     }
 }
